Move non accessible modals to parent modal component
